heatrick/requirements.txt
asyncio==3.4.3
python-kasa==0.7.5